refactored projects section and social links
