+++ src/components/ProjectSection.jsx
@@ -7,6 +7,7 @@ import {
   CardMedia,
   CardContent,
   Grid,
+  Button,
 } from '@mui/material';
 import Projects from '../data/Projects';
 
@@ -18,27 +19,33 @@ export default function ProjectSection() {
       </Typography>
 
       <Grid container spacing={4} justifyContent="center" alignItems="stretch">
-        {Object.keys(Projects).map((project) => (
-          <Grid key={project} item xs={11} sm={6} md={4}>
-            <Card>
-              <CardActionArea>
-                <CardMedia
-                  component="img"
-                  height="150"
-                  src={Projects[project].mainImg}
-                  alt="placeholder image"
-                />
-                <CardContent>
-                  <Typography textAlign="left">{project}</Typography>
-                  <Typography textAlign="left">
-                    {Projects[project].overview}
-                  </Typography>
-                </CardContent>
-              </CardActionArea>
-            </Card>
-          </Grid>
-        ))}
+        {Object.keys(Projects)
+          .slice(0, 6)
+          .map((project) => (
+            <Grid key={project} item xs={11} sm={6} md={4}>
+              <Card>
+                <CardActionArea>
+                  <CardMedia
+                    component="img"
+                    height="150"
+                    src={Projects[project].mainImg}
+                    alt="placeholder image"
+                  />
+                  <CardContent>
+                    <Typography textAlign="left">{project}</Typography>
+                    <Typography textAlign="left">
+                      {Projects[project].overview}
+                    </Typography>
+                  </CardContent>
+                </CardActionArea>
+              </Card>
+            </Grid>
+          ))}
       </Grid>
+
+      <Button color="secondary" variant="contained" sx={{ mt: 2 }}>
+        View More
+      </Button>
     </Box>
   );
 }
